Minimal SSR fix.
See e.g. https://github.com/webpack/webpack/issues/6522 /
https://github.com/webpack/webpack/issues/5843
and the docs at
https://webpack.js.org/configuration/output/#outputglobalobject

This was broken by the Webpack 4 upgrade in #273.
This fix effectively reinstates Webpack 3 behaviour in the browser.

Note the component still doesn't really support SSR, but
its calls to Browser APIs are in componentDidMount and similar
so are delayed until client side.